+++ NEWS.md
@@ -5,6 +5,7 @@ ethos 0.2.1 pre-release
 =======================
 
 - Adds builtin list operators `eo::list_diff` (difference) and `eo::list_inter` (intersection)
+- Adds support for explicit conclusions to proof rules via `:conclusion-explicit`.
 
 ethos 0.2.0
 ===========
@@ -33,7 +34,6 @@ This release of Ethos is associated with the 1.3.0 release of the SMT solver cvc
 - The option `--binder-fresh`, which specified for fresh variables to be constructed when parsing binders, has been removed.
 - Programs and oracles now are explicitly required to have at least one argument.
 - Remove support for the explicit parameter annotation `eo::_`, which was used to provide annotations for implicit arguments to parameterized constants.
-- Adds support for explicit conclusions to proof rules via `:conclusion-explicit`.
 - Programs are now recommended to use the attribute `:signature` to specify the argument and return types.
 
 ethos 0.1.1